fix(frontend): fix no-tracing mode while loading from api
