+++ web/src/pages/Home/Home.tsx
@@ -3,17 +3,20 @@ import withAnalytics from 'components/WithAnalytics/WithAnalytics';
 import {useDataStoreConfig} from 'providers/DataStoreConfig/DataStoreConfig.provider';
 import CreateTransactionProvider from 'providers/CreateTransaction';
 import CreateTestProvider from 'providers/CreateTest';
-import ConfigCTA from './ConfigCTA';
 import Content from './Content';
 
 const Home = () => {
-  const {shouldDisplayConfigSetup, skipConfigSetup} = useDataStoreConfig();
+  const {isLoading, shouldDisplayConfigSetup, skipConfigSetup} = useDataStoreConfig();
 
   return (
     <Layout hasMenu>
       <CreateTransactionProvider>
         <CreateTestProvider>
-          {shouldDisplayConfigSetup ? <ConfigCTA onSkip={skipConfigSetup} /> : <Content />}
+          <Content
+            isLoading={isLoading}
+            shouldDisplayConfigSetup={shouldDisplayConfigSetup}
+            skipConfigSetup={skipConfigSetup}
+          />
         </CreateTestProvider>
       </CreateTransactionProvider>
     </Layout>
